+++ server/src/rtsim/mod.rs
@@ -20,7 +20,7 @@ use slab::Slab;
 use specs::{DispatcherBuilder, WorldExt};
 use vek::*;
 
-pub use self::entity::Entity;
+pub use self::entity::{Brain, Entity, RtSimEntityKind};
 
 pub struct RtSim {
     tick: u64,
@@ -104,7 +104,11 @@ pub fn add_server_systems(dispatch_builder: &mut DispatcherBuilder) {
     ]);
 }
 
-pub fn init(state: &mut State, #[cfg(feature = "worldgen")] world: &world::World) {
+pub fn init(
+    state: &mut State,
+    #[cfg(feature = "worldgen")] world: &world::World,
+    #[cfg(feature = "worldgen")] index: world::IndexRef,
+) {
     #[cfg(feature = "worldgen")]
     let mut rtsim = RtSim::new(world.sim().get_size());
     #[cfg(not(feature = "worldgen"))]
@@ -113,22 +117,69 @@ pub fn init(state: &mut State, #[cfg(feature = "worldgen")] world: &world::World
     // TODO: Determine number of rtsim entities based on things like initial site
     // populations rather than world size
     #[cfg(feature = "worldgen")]
-    for _ in 0..world.sim().get_size().product() / 400 {
-        let pos = rtsim
-            .chunks
-            .size()
-            .map2(TerrainChunk::RECT_SIZE, |sz, chunk_sz| {
-                thread_rng().gen_range(0..sz * chunk_sz) as i32
-            });
+    {
+        for _ in 0..world.sim().get_size().product() / 400 {
+            let pos = rtsim
+                .chunks
+                .size()
+                .map2(TerrainChunk::RECT_SIZE, |sz, chunk_sz| {
+                    thread_rng().gen_range(0..sz * chunk_sz) as i32
+                });
 
-        rtsim.entities.insert(Entity {
-            is_loaded: false,
-            pos: Vec3::from(pos.map(|e| e as f32)),
-            seed: thread_rng().gen(),
-            controller: RtSimController::default(),
-            last_time_ticked: 0.0,
-            brain: Default::default(),
-        });
+            rtsim.entities.insert(Entity {
+                is_loaded: false,
+                pos: Vec3::from(pos.map(|e| e as f32)),
+                seed: thread_rng().gen(),
+                controller: RtSimController::default(),
+                last_time_ticked: 0.0,
+                kind: RtSimEntityKind::Random,
+                brain: Default::default(),
+            });
+        }
+        for (site_id, site) in world
+            .civs()
+            .sites
+            .iter()
+            .filter_map(|(site_id, site)| site.site_tmp.map(|id| (site_id, &index.sites[id])))
+        {
+            use world::site::SiteKind;
+            #[allow(clippy::single_match)]
+            match &site.kind {
+                #[allow(clippy::single_match)]
+                SiteKind::Dungeon(dungeon) => match dungeon.dungeon_difficulty() {
+                    Some(5) => {
+                        let pos = site.get_origin();
+                        if let Some(nearest_village) = world
+                            .civs()
+                            .sites
+                            .iter()
+                            .filter(|s| s.1.is_settlement())
+                            .min_by_key(|(_, site)| {
+                                let wpos = site.center * TerrainChunk::RECT_SIZE.map(|e| e as i32);
+                                wpos.map(|e| e as f32)
+                                    .distance_squared(pos.map(|x| x as f32))
+                                    as u32
+                            })
+                            .map(|(id, _)| id)
+                        {
+                            for _ in 0..25 {
+                                rtsim.entities.insert(Entity {
+                                    is_loaded: false,
+                                    pos: Vec3::from(pos.map(|e| e as f32)),
+                                    seed: thread_rng().gen(),
+                                    controller: RtSimController::default(),
+                                    last_time_ticked: 0.0,
+                                    kind: RtSimEntityKind::Cultist,
+                                    brain: Brain::raid(site_id, nearest_village),
+                                });
+                            }
+                        }
+                    },
+                    _ => {},
+                },
+                _ => {},
+            }
+        }
     }
 
     state.ecs_mut().insert(rtsim);